Add files to initialize folder
This update provides the addition of three script files that were used to (1) convert a folder containing .cif structure files to .cel structure files for image simulation in the Dr. Probe software, (2) to plot an image file in the .dat format, and (3) to rotate a supported Pt nanoparticle about arbitrary angles and to delete the atoms that are embedded in the support.